Create FaceTrackingRenderer2D.cpp
This code is part of the Face Tracking(c++) demo of the F200 Intel RealSense camera. For my project I only modified the file "FaceTrackingRenderer2D.cpp" and added the code between "//Author: ...." and ·"//End of section" You will need to add this cpp to the demo project in order to function.